requirements.txt
Flask
requests
google-generativeai
PyMuPDF
